[#3] Fixed path issue
Initially we used /opt/akvo-dash but then moved to /app inside the
container since it’s simpler if we only run one thing in a container to
always have that in /app. Had forgot to swap to use that in the
docker-compose.yml file so what was executed was was was copied onto
the container when it was build and not the local mount point.