[#130] Test the Tardis
Three tests to test the basics of the triggers that populate the history schema tables.

Test that that insertion of a record in the parent table creates a record in the history table, and that it's selectable by a timestamp later than when the record was created.

Test that update of the record created in the first test creates a new record in the history table. Test that this record holds the updated information. Test that there are now 2 records in the history table.

Test that the deletion of the original record doesn't add a record in the history. Test that there is no "current" record in the history.