Issue/260 domain logic (#351)
* [#260]

Move the domain logic for sharing visualisation and dashboards to it’s
own /lib/ namespace. Also moved implementation to the share-impl
namespace in a try to get a easy parsable API in the non -impl
namespace.

* [#260] Explicit require

* [#260] Split dashboard domain logic

* [#260] Split domain logic for dataset

* [#260] Split Flow domain logic

* [#260] Split job execution domain logic

* [#260] Split public endpoint domain logic

* [#260] Split visualisation domain logic

Also made the library use the domain logic.

* [#260] Fixed issue with no tenant-conn in function call

* [#260] Fixed two issues

- Fixed issue with creating visualisations
- Fixed issue with sharing items

* [#260] Remove comments

* [#260]

Move the domain logic for sharing visualisation and dashboards to it’s
own /lib/ namespace. Also moved implementation to the share-impl
namespace in a try to get a easy parsable API in the non -impl
namespace.

* [#260] Explicit require

* [#260] Split dashboard domain logic

* [#260] Split domain logic for dataset

* [#260] Split Flow domain logic

* [#260] Split job execution domain logic

* [#260] Split public endpoint domain logic

* [#260] Split visualisation domain logic

Also made the library use the domain logic.

* [#260] Fixed issue with no tenant-conn in function call

* [#260] Fixed two issues

- Fixed issue with creating visualisations
- Fixed issue with sharing items

* [#260] Remove comments

* [#260] Clean up visualisation

Also fixed bug with library

* [#260] Cleanup dataset domain logic

Also fixed issue with seed data being test data.

* [#260] Clean up of shares + ns declarations

* [#260] Fixed bug introduced with fn name change

* [#260] Moved hugsql files to lib/ ns

Since the sql files should live with the domain logic.

* [#260] Clean up dashboard domain logic

Also fixed various issues in related tests.

* [#260] Small fixes from PR review feedback

* [#260] Domain logic return ring responses

Since the response depends on domain logic we moved the reponse generation.

* [#260] Fixed lint issues