[#455] Add namespacing to keycloak groups
Since we live in Keycloak with other services we should not steal top
level groups but namespace as per the roles. We added a akvo/lumen/
prefix. Also updated the dev Keycloak json blob to match this.

Since we had to recreate the keycloak json blob the clients got named
as they used to be “akvo-lumen”. Then we don’t have to config a new
client but can reuse the current one.

Also fixed some issues with configs.