Using the Postgis alpine image to see if we avoid the intermittent build failures in CI
It seems that Docker Hub tags are not tags in the VCS sense, so images are rebuild.

Looking at the image that we are using: https://hub.docker.com/r/mdillon/postgis/builds/ it seems that 9.6 worked 16 days ago, but the last time that was built failed.

We guess that some Docker CDN have the 9.6 image while other do not.
To fix the issue, changing the image to 9.6-alpine, which had the last build ok.