[#1161] Run keytool just if the certificate is not already imported
When restarting the BE container in dev mode, keytool is always run, which causes keytool to exit with a non-0 status.

We just added "set -e" to the script, so the BE script was exiting the second time that it was started.

Changed the script so that the import just happens if the certificate is not in the store