[#843] Set all monitoring urls to "unknown"
There is no way to get the full compojure matched path, so instead of
having thousands of urls (due to the REST urls), prefer to have just one
stat, so at least we can see the traffic and error rates.

Compojure has a wrap-routes function to do this, but it doesnt work as
we need when using compojure/context. When using context, we can just
capture the latest bit of the matched api. For example:

(context "/visualization/:id"
  (GET "/meta"...

We are just able to get the "/meta" part, when what we actually want is
"/visualization/:id/meta".