Use string type as transformation.op value
Keyword aren't jsonables in db (postgres/json) nor client
side (javascript/json)

So far we change `op` value to keyword when backend tries to
process/dispatch the transformation client request, that force
implementation to manage the same data with different types, firstly
as string, later as keyword, finally as string.

Also the approach is not the dafault one. Dispatching aggregation is based on
strings, also dispatching import process is based on strings too

So this commit will unify the dispatching approach to be always based
on strings