[#741] Pivot table full filter of values
It's possible to filter out all rows that is needed as categories for
crosstab(text source_sql, text category_sql) in postgres. According to Postgres docs
this is not valid:
"category_sql is a SQL statement that produces the set of categories. This
statement must return only one column. It must produce at least one row,
or an error will be generated" (https://www.postgresql.org/docs/11/tablefunc.html).

Hence when there is no categores we return an empty table.