[#2127] Enhance js code checking via AST parsing instead of strings
Nashorn contains a `parser.js` library following the same Esprima API.
We can do the same checks in the backend so we allow filter|map|reduce
but we limit all the JS code to *one* expression.

There is also the Nashorn Parser API [1] at Java level, but it means dealing
with Java interop with the visitor pattern, etc.

[1] https://docs.oracle.com/javase/9/docs/api/jdk/nashorn/api/tree/Parser.html