[#2127] Use `nashorn:parser.js` for code parsing
In the previous attempt the code parsing the JS expression was still
using a "cached" version of esprima.js (thanks to REPL driven
development).

It's possible to use esprima.js as dependency and use `load()` to
have the `esprima.parse()` available. The problem with this approach
is that is way slower than use the the `nashorn:parser.js`.
Nashorn needs to load the file, parse and compile the code.

The JSON representation of both parsers is the same. Instead of
dealing with "live" JS Objects, we want to deal with Clojure data
structures. We can parse the code, "stringify" the result and parse it
again in Clojure side and is faster than using `esprima` as
dependency.