[#2127] Automatic JSON parsing to simplify JS expressions
We're storing the JSON blob as a string. In order to make the user enter less code
we're helping by automatic parsing `rqg` columns.

Assuming an `age` columns with values: "[10, 5, 15, 20]"
Instead of:

    JSON.parse(row["age"]).filter(a => a > 10)

The user can use the row["age"] directly

    row["age"].filter(a => a > 10)

Or a more complex one

    JSON.parse(row["age"]).reduce((a,b) => (a+b), 0) / JSON.parse(row["age"]).length

with automatic parsing:

    row["age"]).reduce((a,b) => (a+b), 0) / row["age"].length

We tried to parse the JSON blob on the Clojure side, this works but
the resulting datastructure is not JS array `[object Array]`, and it
doesn't have the `.filter()` `.map()` `.reduce()` function.

The workaround is to add to `row` the list of columns with a `rqg`
type, and to the `JSON.parse` before executing the user's code.