[#2127] When `rgq` flag is on, add `ns` and `repeatable` keys
The client needs to know which data groups are repeatable. For that,
we client can rely on the metadata in the column definition.

After importing a Flow form, we store as part of the columns
definition the `ns` and `repeatable` keys in the columns and expose
them as part of the endpoint `/dataset/<id>/groups`.