[#3045] Hide `Instance id` column from non-metadata groups
The `Instance id` column is a special column that serves as link
(internally a foreign key) for all data groups to the parent one
(metadata). The user must not be able to transform this
column. Initially we're going to hide it.

Since the column already has a property `hidden` we just use it to
include or not the column.