[#3045] Hide `Instance id` in the backend
We have explored the idea of hiding the the `Instance id` for
non-metadata groups in the client. The result is that some code uses
the Redux store (a central place for state) but other parts, like the
Grid, uses the direct response from the XHR call.

Hiding only visually in the Grid and Dropdowns makes us wonder how the
grid will behave if the server is responding more data than the
visible one. How this will play out when transformations are added.

For now it seems *safer* to filter the `hidden` columns in the
backend.